refactor: updating the typing and giving typing to balances handler
